% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure-report.R
\name{configure_report}
\alias{configure_report}
\title{Configure report template}
\usage{
configure_report(
  add_introduction = TRUE,
  add_plot_introduction = TRUE,
  introduction_args = list(),
  plot_introduction_args = list(),
  add_summary = TRUE,
  add_plot_summary = TRUE,
  summary_args = list(),
  plot_summary_args = list(),
  global_ggtheme = quote(theme_gray()),
  global_theme_config = list()
)
}
\arguments{
\item{add_introduction}{add \link{introduction}? Default is \code{TRUE}.}

\item{add_plot_introduction}{add \link{plot_introduction}? Default is \code{TRUE}.}

\item{introduction_args}{arguments to be passed to \link{introduction}. Default is \code{list()}.}

\item{plot_introduction_args}{arguments to be passed to \link{plot_introduction}. Default is \code{list()}.}

\item{add_summary}{add \link{summary}? Default is \code{TRUE}.}

\item{add_plot_summary}{add \link{plot_summary}? Default is \code{TRUE}.}

\item{summary_args}{arguments to be passed to \link{summary}. Default is \code{list()}.}

\item{plot_summary_args}{arguments to be passed to \link{plot_summary}. Default is \code{list()}.}

\item{global_ggtheme}{global setting for \link{theme}. Default is \code{quote(theme_gray())}.}

\item{global_theme_config}{global setting for \link{theme}. Default is \code{list()}.}
}
\description{
This function configures the content of the to-be-generated data profiling report.
}
\note{
Individual settings will overwrite global settings. For example: if \code{plot_introduction_args} has \code{ggtheme} set to \code{theme_light()} while \code{global_ggtheme} is set to \code{theme_gray()}, \code{theme_light()} will be used.

When setting global themes using \code{global_ggtheme}, please pass an unevaluated call to the theme function, e.g., \code{quote(theme_light())}.
}
\examples{
## Get default configuration
configure_report()

## Set global theme
configure_report(global_ggtheme = quote(theme_light(base_size = 20L)))
}
\seealso{
\link{create_report}
}
\keyword{configure_report}
