% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_fields.R
\name{counts_fields}
\alias{counts_fields}
\title{Group_by and count the occurances of a group in a data set}
\usage{
counts_fields(data, field_name)
}
\arguments{
\item{data}{The tibble to count groups}

\item{field_name}{The name of the field to group and count}
}
\description{
This takes a tibble and a field name and does a group_by that field and
counts the results
}
\details{
Group a dataset by a field and provide the counts of each group
}
\examples{
counts_fields(practice_data, practice_area)

}
