% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_excel_data.R
\name{get_excel_data}
\alias{get_excel_data}
\title{Get data from an Excel file}
\usage{
get_excel_data(file, sheet_name = "Sheet1", skip_rows = 0)
}
\arguments{
\item{file}{The relative path to the Excel data file.}

\item{sheet_name}{The name of the particular worksheet from which we wish to
get data.  The default is Sheet1.}

\item{skip_rows}{The number of rows in the sheet to skip before processing
the file.  The default is 0. This allows the user to deal with files that have
rows that contain data that we do not wish to use in our analysis.}
}
\description{
A wrapper around the readxl::read_excel function to get data from an Excel
file.  It is a convenience function for using targets.  As most of the data
cleaning functions that have been moved to the clean_* set of functions, this
may not need to exist anymore, but it is left for heritage purposes.
}
\details{
Retrieve the data stored in an Excel file
}
