% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemapr.R
\docType{package}
\name{rolemapr-package}
\alias{rolemapr-package}
\alias{rolemapr}
\alias{Rolemapr}
\alias{Rolemapr-package}
\title{Role Mapper}
\description{
A package to produce roles and responsibility maps.
}
