% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_report.R
\name{configure_report}
\alias{configure_report}
\title{Configure report template}
\usage{
configure_report(
  add_introduce = TRUE,
  introduce_args = list(),
  global_ggtheme = quote(theme_gray()),
  global_theme_config = list()
)
}
\arguments{
\item{add_introduce}{add \link{introduce}? Default is \code{TRUE}.}

\item{introduce_args}{arguments to be passed to \link{introduce}. Default is \code{list()}.}

\item{global_ggtheme}{global setting for theme.Default is \code{quote(theme_gray())}.}

\item{global_theme_config}{global setting for theme. Default is \code{list()}.}
}
\description{
This function configures the content of the to-be-generated data profiling report.
}
\note{
Individual settings will overwrite global settings. For example: if \code{intro_args} has \code{ggtheme} set to \code{theme_light()} while \code{global_ggtheme} is set to \code{theme_gray()}, \code{theme_light()} will be used.

When setting global themes using \code{global_ggtheme}, please pass an unevaluated call to the theme function, e.g., \code{quote(theme_light())}.
}
\examples{
## Get default configuration
configure_report()

## Set global theme
configure_report(global_ggtheme = quote(theme_light(base_size = 20L)))

}
\seealso{
\link{create_report}
}
\keyword{configure_report}
