% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemapr-package.R
\docType{package}
\name{rolemapr-package}
\alias{rolemapr}
\alias{rolemapr-package}
\title{rolemapr: A package to produce roles and responsibility maps}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whjelmar/rolemapr}
  \item \url{https://whjelmar.github.io/rolemapr/}
  \item Report bugs at \url{https://github.com/whjelmar/rolemapr/issues}
}

}
\author{
\strong{Maintainer}: Walter Hjelmar \email{walter@hjelmar.com} (\href{https://orcid.org/0000-0002-0625-2933}{ORCID})

}
\keyword{internal}
